Replace deprecated FileFinder.find_module (Python 3.12)
The find_module function has been deprecated for a long time and
it has been removed in Python3.12. To keep the backward compatibility,
we introduce compat.load_module() function that is compatible with
Python2.7 -> Python3.12 now. Replacing all previous load of module
in code by this function.

Similar applies for the pkgutil.get_loader()

Co-authored-by: Matej Matuska <mmatuska@redhat.com>