snactor(run): allow using leapp's execution contexts
Introduce the 'LEAPP_DEBUG_PRESERVE_CONTEXT' environmental variable.
When the variable is set to 1 and the environment has
'LEAPP_EXECUTION_ID' set, the 'LEAPP_EXECUTION_ID' is not overwritten
with snactor's execution ID. This allows the developer to run actors
in the same fashion as if the actor was run during the last leapp's
execution, thus, avoiding to rerun the entire upgrade process.
This, naturally, does not help when the actor has outside dependencies,
i.e., it requires the filesystem to be set up in a specific way (e.g.
a target container) must be present, e.t.c.