Separate leapp version to __version__.py
Originally the version of the leapp module has been defined inside
`leapp/__init__.py` module which has been loaded when building
the leapp python module/package. Part of the `leapp/__init__.py` file
is also load and application of workarounds right now, which is unsafe
and adding new workarounds can lead to crash of the build process as
the required runtime python modules do not have to be installed on
the system that time.

Moving the definition into the separate `leapp/__version__.py` file
which could be used safely during the build time. Note that nowadays
suggested python way requires Python 3.8+ but this project is going
to support Python3.6 for another years, so we go this old way still.