CI:packit: update e2e tests to reflect current IPU paths
The current source version used in e2e tests is 8.6 and the target version is 9.0, which are unsupported. This patch bumps the source/target versions used in the CI to source=8.10 and target=9.4. There is also a job with target=9.6, which is expected to be failing until 9.6 is released.

Also rhel 7 & python2 is not supported anymore. Dropping these obsoleted tests.