Fix CLI: allow to set falsy values as default
When specifying an option for a leapp CLI command using @add_option
decorator, it has not been possible to specify a "falsy" value
(or let's say empty) as `default` e.g. empty string, empty list, etc.
Start to check NoneType for the `default` parameter properly to
allow it to set this.

Bump leapp-framework to 6.1.

Note:
The similar problem is for `choices` and `metavar` parameters.
In case of `metavar` it's however ok to ignore empty string.
In case of empty `choices` it would lead to the same evaluation as
in case of NoneType (any input is ok). So I that the `default` is the
only parameter that needs to be checked properly.

JIRA: RHEL-86206